+++ integer/CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## Unreleased
+
+- Implement `TryFrom<UBig>` for `f32`/`f64`.
+
 ## 0.4.1
 
-- Implemented `AbsEq` and `AbsOrd` for `UBig` and `IBig`.
+- Implement `AbsEq` and `AbsOrd` for `UBig` and `IBig`.
 - Add `UBig::from_static_words` and `IBig::from_static_words` (both Rust 1.64+) to support the `static_ubig!` and `static_ibig!` macros.
 - Add `is_multiple_of` and `is_multiple_of_const` (Rust 1.64+) for `UBig`/`IBig`
 - Constify `trailing_zeros` and `trailing_ones` of `UBig`/`IBig` (Rust 1.64+).